Fix: Corrected a bug which made the importing of races from other devices inconsistent.
Virtual Race is designed to search the iCloud servers for races the user is in that are not currently stored on the device they are using. If it finds any, Virtual Race will add those races to the device. Changes in swift syntax made this feature inconsistent but have now been corrected.